+++ temp_cata/BetterBlizzPlates.lua
@@ -9,7 +9,7 @@ LSM:Register("font", "Yanone (BBP)", [[Interface\Addons\BetterBlizzPlates\media\
 LSM:Register("font", "Prototype", [[Interface\Addons\BetterBlizzPlates\media\Prototype.ttf]])
 
 local addonVersion = "1.00" --too afraid to to touch for now
-local addonUpdates = "1.6.8"
+local addonUpdates = "1.6.8b"
 local sendUpdate = false
 BBP.VersionNumber = addonUpdates
 local _, playerClass